% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cbind_rcpp}
\alias{cbind_rcpp}
\title{cbind_rcpp}
\arguments{
\item{A}{es una matriz numerica de dimención nxm}

\item{B}{es un vector de largo n}
}
\value{
matriz con los datos del vector insertados en una nueva columna
}
\description{
Esta fución une un vector y una matriz siempre y cuando el largo
del vector sea igual a la candidad de filas que tiene la matriz.
}
\details{
[[Rcpp::export]]
}
\examples{
matriz <- matrix(round(runif(9, 1, 10)), nrow = 3, ncol = 3)
vector <- c(1:3)
set.seed(123)
result <- cbind_rcpp(matriz,vector)
result
}
