% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{maxMin_cpp}
\alias{maxMin_cpp}
\title{maxMin_cpp}
\arguments{
\item{A}{es una matriz numerica de dimención nxn}

\item{B}{es una matriz numerica de dimención nxn}
}
\value{
regresa una matriz numerica de dimencion nxn
}
\description{
Esta funcion se encarga de buscar el maximo de los minimos de ambas matrices
para devolver una matriz diferente con los valores encontrados.
}
\details{
[[Rcpp::export]]
}
\examples{
A <- matrix(round(runif(n, min = 0, max = 1), 1), nrow = 3, ncol = 3)
B <- matrix(round(runif(n, min = 0, max = 1), 1), nrow = 3, ncol = 3)
diag(A) <- 1
diag(B) <- 1
set.seed(123)
result <- maxMin_cpp(A,B)
result
}
