% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{whichMax}
\alias{whichMax}
\title{whichMax}
\arguments{
\item{A}{es una matriz numerica de dimención nxm}
}
\value{
retorna un data.frame
}
\description{
Esta funcion devuelve un data.frame en donde se encuentran las 
coordenadas de los valores maximos ubicados dentro de la matriz.
}
\details{
[[Rcpp::export]]
}
\examples{
matriz <- matrix(round(runif(9, 1, 10)), nrow = 3, ncol = 3)
set.seed(123)
result <- whichMax(matriz)
result
}
